test-mcps/postgres-mcp/pyproject.toml
[project]
name = "langsight-postgres-mcp"
version = "0.1.0"
description = "PostgreSQL MCP server for LangSight testing"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]

[project.scripts]
postgres-mcp = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
